# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinical Review Copilot API (P-194)",
    "description": "Backend API for Clinical Review Copilot platform — FHIR R4 & PDF/OCR Ingestion",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_api_v1_patients_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/timeline": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Timeline",
        "operationId": "get_timeline_api_v1_patients__patient_id__timeline_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/trends": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Trends",
        "operationId": "get_trends_api_v1_patients__patient_id__trends_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mã LOINC xét nghiệm",
              "title": "Code"
            },
            "description": "Mã LOINC xét nghiệm"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/drug-interactions": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Drug Interactions",
        "operationId": "get_drug_interactions_api_v1_patients__patient_id__drug_interactions_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrugInteractionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/memory": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient Memory",
        "operationId": "get_patient_memory_api_v1_patients__patient_id__memory_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMemory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "Delete Patient",
        "operationId": "delete_patient_api_v1_patients__patient_id__delete",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/care-plan": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Generate Patient Care Plan",
        "description": "Generate dynamic patient home care plan using Clinical LLM Agent & Medical RAG.",
        "operationId": "generate_patient_care_plan_api_v1_patients__patient_id__care_plan_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DrugInteractionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DrugInteractionFlag"
            },
            "type": "array",
            "title": "Items"
          },
          "data_watermark": {
            "type": "string",
            "title": "Data Watermark"
          }
        },
        "type": "object",
        "required": [
          "items",
          "data_watermark"
        ],
        "title": "DrugInteractionListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PatientListResponse"
      },
      "PatientMemory": {
        "properties": {
          "memory_version_id": {
            "type": "string",
            "title": "Memory Version Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "source_review_version_id": {
            "type": "string",
            "title": "Source Review Version Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By"
          },
          "approved_at": {
            "type": "string",
            "title": "Approved At"
          }
        },
        "type": "object",
        "required": [
          "memory_version_id",
          "version",
          "patient_id",
          "source_review_version_id",
          "approved_by",
          "approved_at"
        ],
        "title": "PatientMemory"
      },
      "PatientSummary": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "pseudonym": {
            "type": "string",
            "title": "Pseudonym"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "sex": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "unknown"
            ],
            "title": "Sex"
          },
          "primary_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Condition"
          },
          "last_encounter_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Encounter At"
          },
          "latest_data_watermark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Data Watermark"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "pseudonym",
          "sex"
        ],
        "title": "PatientSummary"
      },
...
      "TimelineEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "encounter",
              "observation",
              "medication",
              "condition",
              "allergy",
              "note"
            ],
            "title": "Event Type"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "citations": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentCitation"
                },
                {
                  "$ref": "#/components/schemas/FhirCitation"
                },
                {
                  "$ref": "#/components/schemas/RecordCitation"
                }
              ]
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "occurred_at",
          "title",
          "summary"
        ],
        "title": "TimelineEvent"
      },
      "TimelineResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Items"
          },
          "data_watermark": {
            "type": "string",
            "title": "Data Watermark"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "data_watermark",
          "page",
          "page_size",
          "total"
        ],
        "title": "TimelineResponse"
      },
...
      "TrendResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "profile_version": {
            "type": "string",
            "title": "Profile Version",
            "default": "type_2_diabetes@1.0.0"
          },
          "data_watermark": {
            "type": "string",
            "title": "Data Watermark"
          }
        },
        "type": "object",
        "required": [
          "code",
          "display",
          "unit",
          "points",
          "data_watermark"
        ],
        "title": "TrendResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}